sync
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@4689 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ DCMView.h
@@ -327,7 +327,7 @@ typedef enum {DCMViewTextAlignLeft, DCMViewTextAlignCenter, DCMViewTextAlignRigh
 - (void) initFont;
 - (void) gClickCountSetReset;
 - (long) indexForPix: (long) pixIndex; // Return the index into fileList that coresponds to the index in pixList
-- (long) findPlaneAndPoint:(float*) pt :(float*) location;
+- (int) findPlaneAndPoint:(float*) pt :(float*) location;
 - (unsigned char*) getRawPixels:(long*) width :(long*) height :(long*) spp :(long*) bpp :(BOOL) screenCapture :(BOOL) force8bits;
 
 - (unsigned char*) getRawPixelsWidth:(long*) width height:(long*) height spp:(long*) spp bpp:(long*) bpp screenCapture:(BOOL) screenCapture force8bits:(BOOL) force8bits removeGraphical:(BOOL) removeGraphical squarePixels:(BOOL) squarePixels allTiles:(BOOL) allTiles allowSmartCropping:(BOOL) allowSmartCropping origin:(float*) imOrigin spacing:(float*) imSpacing;
